Adding the search suggestion master CSV
This csv will be used in the initialization process of the search suggestion Trie at the initialization of the search engine. This will be written at the start and read for the individual search suggestion operations. If new search suggestions are to be created from searches during runtime, those searches should be appended as new lines to suggestionSeeds.txt.