Adding CSV prep program
This program uses an api called requests to pull the most common auto suggestions to searchs from google. I feed the suggestions seeds from the suggestionSeeds.txt, and then write those to the CSV. Include the python file locally to the driver program then run prepSuggestionMaster() to initialize the CSV for a given instance of the search engine.